requirement.txt
anyio==4.3.0
astor==0.8.1
attrdict==2.0.1
Babel==2.15.0
bce-python-sdk==0.9.10
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
contourpy==1.2.1
cssselect==1.2.0
cssutils==2.11.0
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
einops==0.8.0
et-xmlfile==1.1.0
filelock==3.14.0
fire==0.6.0
Flask==3.0.3
flask-babel==4.0.0
fonttools==4.51.0
fsspec==2024.5.0
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.7
imageio==2.34.1
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.4.1
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==24.0
paddleocr==2.7.3
paddlepaddle==2.6.1
pandas==2.2.2
pdf2docx==0.5.8
pillow==10.3.0
premailer==3.10.0
protobuf==5.26.1
psutil==5.9.8
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pycryptodome==3.20.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
rapidfuzz==3.9.1
rarfile==4.2
regex==2024.5.15
requests==2.32.2
safetensors==0.4.3
scikit-image==0.23.2
scipy==1.13.0
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==68.0.0
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sympy==1.12
termcolor==2.4.0
thop==0.1.1.post2209072238
tifffile==2024.5.10
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.41.0
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.18
urllib3==2.2.1
visualdl==2.5.3
Werkzeug==3.0.3
wheel==0.41.2

